+++ lib/src/utils/utils.dart
@@ -1,5 +1,6 @@
 export 'apply_aspect_ratio.dart';
 export 'calculate_axis_path.dart';
+export 'calculate_radius_path.dart';
 export 'calculate_rounded_path.dart';
 export 'degrees_to_radians.dart';
 export 'flatten_segments.dart';
@@ -7,5 +8,6 @@ export 'get_arc_angle.dart';
 export 'get_point_on_circle.dart';
 export 'get_sagitta.dart';
 export 'lerp_double.dart';
+export 'null_tween.dart';
 export 'rotate_over_origin.dart';
 export 'rounded_poly.dart';