+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-## 0.4.0-dev.5
+## 0.4.0-dev.6
 - Added factory constructors _GaugeProgressBar_ and _GaugePointer_ classes.
 - Fixed the `radius` property of *RadialGauge* and *AnimatedRadialGauge* widgets.
 - Added implicit `radius` animation for the *AnimatedRadialGauge* widget.
 - Added `cornerRadius` argument for *GaugeAxisStyle* and *GaugeSegment* classes.
 - Fixed invalid segment spacing.
 - Fixed the basic progress bar for the 360-degree axis.
+- Fixed errors that occur when using the *AnimatedRadialGauge* widget without the *radius* property provided.
+- Added default style for *RadialGauge* and *AnimatedRadialGauge* widgets to make it easier to get started with the widget.
 - Minor bug fixes.
 
 ## 0.3.3